Add a ChuckedDBWrapper and chunked options
This patch adds some cli options to the migrate sub command:
  --chunked - To enable chunking mode (defaults to False)
  --chunk-size - to define the size of the chunks (defaults to 10000)
  --status-dir - To define a location so save migration status to.
                 (default /tmp/psql2mysql/)

And all this is implemented as a new DB wrapper. Chunking will
work if a table has a single primary key, if the primary key is
an int or varchar. However, varchars (UUIDs) chunking only works
if there is also a `created_at` field, that we can use for progress.
If none of that is possible, it'll fall back to entire table migrations.