Raise Exception on precheck failures
This introduces a custom RuntimeError Exception that is raised if
something goes wrong during the prechecks or migration.

Addtionally we make sure to use a non-zero exit code if any of the
prechecks fail. And it case of batch processings we stop the batch
now after finding errors in a database.