Add chunked migration using fetchmany
Currently the tool migrates whole database tables at a time, this
patch breaks the mirgation of a table in to chunks of rows.

Chunking up the mirgation saves RAM and so allows the tool to migrate
really large databases. As such, chunking is enabled by default.

The default chunk-size is 100000. This can be changed using the
--chunk-size option. Setting the chunk-size to 0 will disable chunked
migration and data will be migrated as a whole table at a time.