Filter out some SQLalchemy warnings
Some of the database schema on the MySQL side contain things that
SQLalchemy is not able to parse correctly during reflection(), causing
it to print out warnings like:
"SAWarning: Unknown schema content: u'  CONSTRAINT `CONSTRAINT_3` CHECK (`multiattach` in (0,1))'"

For our use case this is actually not a problem. So this commits adds
code to filter out those warnings to avoid confusion.